requirements.txt
pika
mysql-connector-python
pyTelegramBotAPI